+++ app/src/main/java/com/dumpit/ffff/CameraShot.java
@@ -31,6 +31,13 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 import com.gun0912.tedpermission.PermissionListener;
 import com.gun0912.tedpermission.TedPermission;
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdSize;
+import com.google.android.gms.ads.AdView;
+import com.google.android.gms.ads.MobileAds;
+import com.google.android.gms.ads.initialization.InitializationStatus;
+import com.google.android.gms.ads.initialization.OnInitializationCompleteListener;
+
 
 import java.io.File;
 import java.io.IOError;
@@ -52,12 +59,25 @@ public class CameraShot extends AppCompatActivity {
     private String imageFilePath;
     private Uri photoUri;
     int points = 0;
+    private AdView mAdview; //애드뷰 변수 선언
+
 
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.camera);
+        MobileAds.initialize(this, new OnInitializationCompleteListener() { //광고 초기화
+            @Override
+            public void onInitializationComplete(InitializationStatus initializationStatus) {
+            }
+        });
+        mAdview = findViewById(R.id.adView); //배너광고 레이아웃 가져오기
+        AdRequest adRequest = new AdRequest.Builder().build();
+        mAdview.loadAd(adRequest);
+        AdView adView = new AdView(this);
+        adView.setAdSize(AdSize.BANNER); //광고 사이즈는 배너 사이즈로 설정
+        adView.setAdUnitId("\n" + "ca-app-pub-3940256099942544/6300978111");
 
         firebaseDatabase = FirebaseDatabase.getInstance();
         databaseReference = firebaseDatabase.getReference();
